Check and block modifying/deleting unvalid user.

+++ src/main/java/com/practice/Skilltest/board/service/impl/BoardServiceImpl.java
@@ -3,15 +3,22 @@ package com.practice.Skilltest.board.service.impl;
 import com.practice.Skilltest.board.dao.BoardDao;
 import com.practice.Skilltest.board.dto.BoardDto;
 import com.practice.Skilltest.board.service.BoardService;
+import com.practice.Skilltest.user.role.UserRoles;
 import org.apache.ibatis.javassist.NotFoundException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Repository;
 
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 
 @Repository
 public class BoardServiceImpl implements BoardService {
 
+    private final Logger logger = LoggerFactory.getLogger(this.getClass());
     private final BoardDao boardDao;
 
     public BoardServiceImpl(BoardDao boardDao) {
@@ -37,7 +44,7 @@ public class BoardServiceImpl implements BoardService {
         }
     }
 
-    //새 게시글 생성
+    //새 게시글
     public long newBoard(BoardDto req) {
 
         boardDao.newBoard(req);
@@ -45,9 +52,9 @@ public class BoardServiceImpl implements BoardService {
         return req.getBoard_id();
     }
 
-    //게시글 수정
+    //게시글 수정 작업
     @Override
-    public boolean modifyBoard(BoardDto req) {
+    public boolean modifyBoard(BoardDto req, String username, Collection<? extends GrantedAuthority> userAuthority) {
 
         Map<String, Object> map = new HashMap<>();
 
@@ -66,8 +73,35 @@ public class BoardServiceImpl implements BoardService {
 
     //게시글 삭제
     @Override
-    public void deleteBoard(long id) {
+    public boolean deleteBoard(long id, String username, Collection<? extends GrantedAuthority> userAuthority) {
+
+        //요청자 유효 확인
+        if(!checkValidModify(id,username,userAuthority)) {return false;}
+
         boardDao.deleteBoard(id);
+        return true;
     }
+
+
+
+    //요청자의 세션 정보가 해당 게시글의 작성자와 동일자 요청//어드민의 요청인지 확인
+    @Override
+    public boolean checkValidModify(long id, String username, Collection<? extends GrantedAuthority> userAuthority) {
+
+        //요청이 해당 게시글의 게시자와 동일한 유저가 요청하였는지 확인
+        if(boardDao.getWriter(id).equals(username)){ return true; }
+
+        logger.info("삭제자가 게시자와 다른 요청");
+        //만약 아니라면 해당 요청이 어드민 요청인지 확인
+        if(userAuthority.contains(new SimpleGrantedAuthority(UserRoles.ADMIN.getValue()))) {
+            
+            logger.info("어드민 게시글 삭제요청");
+            return true;
+        }
+
+        logger.info("삭제자가 게시자와 다르고 어드민이 아닌 요청 무시");
+        return false;
+    }
+
 }
